requirements.txt
# Core
streamlit>=1.35.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0

# Embeddings & Vector Store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Document loaders
pypdf>=4.0.0
unstructured>=0.15.0

# Web Search
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.0
